+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+### Version 5.1.0-alpha10
+
+- Android Exoplayer: Fix track selections. 
+  - Fix assumption of single render in method getTrackRendererIndex. Provide tracks from different renders of same type in onLoad callback
+  - Dispatch only support track in onLoad callback
+  - Fix choosing single track from a group .getFormat(0) in multiple places.
+  - Fix audio deselection
+  - Use setPreferredAudioLanguage/setPreferredTextLanguage when controlling track selection by language or making a default selection.
+  - Fix id used as audio/text track title  
+
 ### Version 5.1.0-alpha9
 
 - Add ARM64 support for windows [#2137](https://github.com/react-native-community/react-native-video/pull/2137)